[fixed]schema

+++ db/migrate/20210207090340_create_messages.rb
@@ -1,8 +1,8 @@
 class CreateMessages < ActiveRecord::Migration[5.2]
   def change
     create_table :messages do |t|
-      t.bigint "user_id", null: false
-      t.bigint "room_id", null: false
+      t.integer "user_id", null: false
+      t.integer "room_id", null: false
       t.string "message", null: false
       t.timestamps
     end
